在开发模式下重启时不重载临时IP白名单

+++ internal/waf/ip_list.go
@@ -45,8 +45,10 @@ func init() {
 
 	// load
 	go func() {
-		_ = SharedIPWhiteList.Load(cacheFile)
-		_ = os.Remove(cacheFile)
+		if !Tea.IsTesting() {
+			_ = SharedIPWhiteList.Load(cacheFile)
+			_ = os.Remove(cacheFile)
+		}
 	}()
 }
